itch: on install automatically open install dialog in itch client
Use the itch://install?game_id=1234 URI format to automatically launch the game install dialog inside the Itch client. `If the game_id` is invalid, the dialog will not open but the client is still launched, which is about the same as the previous behavior. The install dialog is also correctly launched even if the user needs to click through the Itch login process.