[#3203] Add install support for Itchio (#347)
* [#3203] Add install support for Itchio

There exists an Itch.io url to pass as a command line argument for installing. URLs of the form 'itch://install?game_id=###' will prompt the itch client to open the installation dialog box.

* add missing semicolon

I succeded in compiling the extension and loading it locally. But couldn't compile until I fixed this error